RGZ/backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6
flask-jwt-extended==4.4.4
python-dotenv==1.0.0